backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment Variables
python-decouple==3.8

# Real-time features
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# File Processing
PyMuPDF==1.23.21
python-pptx==0.6.23
Pillow==10.2.0

# AI/ML
google-generativeai==0.3.2
requests==2.31.0

# Utilities
pytz==2024.1

